first commit
sorry i was not able to commit earliar.
but till now the codes for changing damages to millions and billions , making a dictionary for all the separate entries by interating through each list, making a new dictionary in which we get the values through years rather than name of the hurricane, all the affected areas and how many hurricanes hitted that area earliar and sorting all the areas according to the number of hurricanes that hit that place are there.

some of the hints are taken from stack overflow and python.org as well as the answer script from codecademy for better understnding